agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Simon",email = "overdosyu@gmail.com"},
    {name = "jinxi97",email = "jinxi_1997@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=1.0.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]

[tool.poetry]
packages = [{include = "agents", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:start"